Added new 64 bit fields nSize and nUndoSize to CBlockFileInfo and renamed previous 32 bit nSize and nUndoSize to nSizeLegacy and nUndoSizeLegacy. Also updated usages of this two fields to 64 bit.

+++ src/blockfileinfostore.cpp
@@ -13,7 +13,7 @@
 std::unique_ptr<CBlockFileInfoStore> pBlockFileInfoStore = std::make_unique<CBlockFileInfoStore>();
 
 void CBlockFileInfoStore::FindNextFileWithEnoughEmptySpace(const Config &config,
-    unsigned int nAddSize, unsigned int& nFile)
+    uint64_t nAddSize, unsigned int& nFile)
 {
     // this while instead of if is here because first commit introduced it
     // and vinfoBlockFile.size() can exceed nLastBlockFile at least in
@@ -71,7 +71,7 @@ std::vector<std::pair<int, const CBlockFileInfo *>> CBlockFileInfoStore::GetAndC
 
 
 bool CBlockFileInfoStore::FindBlockPos(const Config &config, CValidationState &state,
-    CDiskBlockPos &pos, unsigned int nAddSize, unsigned int nHeight,
+    CDiskBlockPos &pos, uint64_t nAddSize, unsigned int nHeight,
     uint64_t nTime, bool& fCheckForPruning, bool fKnown) {
     LOCK(cs_LastBlockFile);
 
@@ -105,9 +105,9 @@ bool CBlockFileInfoStore::FindBlockPos(const Config &config, CValidationState &s
     }
 
     if (!fKnown) {
-        unsigned int nOldChunks =
+        uint64_t nOldChunks =
             (pos.nPos + BLOCKFILE_CHUNK_SIZE - 1) / BLOCKFILE_CHUNK_SIZE;
-        unsigned int nNewChunks =
+        uint64_t nNewChunks =
             (vinfoBlockFile[nFile].nSize + BLOCKFILE_CHUNK_SIZE - 1) /
             BLOCKFILE_CHUNK_SIZE;
         if (nNewChunks > nOldChunks) {
@@ -137,19 +137,19 @@ bool CBlockFileInfoStore::FindBlockPos(const Config &config, CValidationState &s
 }
 
 bool CBlockFileInfoStore::FindUndoPos(CValidationState &state, int nFile, CDiskBlockPos &pos,
-    unsigned int nAddSize, bool& fCheckForPruning) {
+    uint64_t nAddSize, bool& fCheckForPruning) {
     pos.nFile = nFile;
 
     LOCK(cs_LastBlockFile);
 
-    unsigned int nNewSize;
+    uint64_t nNewSize;
     pos.nPos = vinfoBlockFile[nFile].nUndoSize;
     nNewSize = vinfoBlockFile[nFile].nUndoSize += nAddSize;
     setDirtyFileInfo.insert(nFile);
 
-    unsigned int nOldChunks =
+    uint64_t nOldChunks =
         (pos.nPos + UNDOFILE_CHUNK_SIZE - 1) / UNDOFILE_CHUNK_SIZE;
-    unsigned int nNewChunks =
+    uint64_t nNewChunks =
         (nNewSize + UNDOFILE_CHUNK_SIZE - 1) / UNDOFILE_CHUNK_SIZE;
     if (nNewChunks > nOldChunks) {
         if (fPruneMode) {
